+++ src/main/java/net/thep2wking/oedldoedltechnology/init/ModEntities.java
@@ -19,8 +19,7 @@ public class ModEntities {
 
 		ModRegistryHelper.registerEntityWithSpawnEgg(OedldoedlTechnology.MODID, "republican_space_ranger", OedldoedlTechnology.INSTANCE, id++, EntityRepublicanSpaceRanger.class, 80, 3, false, 0x5d8407, 0x333f00);
 	
-		MinecraftServer server = FMLCommonHandler.instance().getMinecraftServerInstance();
-		if (server != null && server.isSinglePlayer() && TechnologyConfig.CONTENT.SPAWN_REPUBLICAN_SPACE_RANGER) {
+		if (TechnologyConfig.CONTENT.SPAWN_REPUBLICAN_SPACE_RANGER) {
 			ModRegistryHelper.registerEntitySpawn(EntityRepublicanSpaceRanger.class, EnumCreatureType.MONSTER, 25, 1, 2, ModBiomeUtil.getEntitySpawningBiomes(EnumCreatureType.MONSTER, EntityZombie.class));
 		}
 	}